+++ CHANGELOG.md
@@ -12,6 +12,7 @@ CHANGELOG
 - add opened file name to application title;
 - avoid "Broken pipe" errors if command-line output is piped and interrupted;
 - fixed error on opening databases with invalid SkypeOut values in account;
+- fixed invalid HTML in export;
 - command-line output encoded on all platforms, to avoid Unicode errors;
 - avoid stack trace on command-line keyboard interrupt;
 - stop worker threads instead of using daemons, to avoid zombie messages in CLI;